requirements.txt
absl-py==0.9.0
astroid==2.4.2
attrs==19.3.0
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
bleach @ file:///home/conda/feedstock_root/build_artifacts/bleach_1588608214987/work
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
gluoncv==0.7.0
google-auth==1.19.2
google-auth-oauthlib==0.4.1
graphviz==0.8.4
grpcio==1.30.0
idna==2.6
importlib-metadata @ file:///D:/bld/importlib-metadata_1593211549301/work
ipykernel @ file:///D:/bld/ipykernel_1595101481730/work/dist/ipykernel-5.3.3-py3-none-any.whl
ipython @ file:///D:/bld/ipython_1593235603364/work
ipython-genutils==0.2.0
isort==4.3.21
jedi @ file:///D:/bld/jedi_1595018491436/work
Jinja2==2.11.2
jsonschema==3.2.0
jupyter-client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1594732094290/work
jupyter-core==4.6.3
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.2.2
mccabe==0.6.1
mistune==0.8.4
mxboard==0.1.0
mxnet-cu101==1.5.0
nbconvert==5.6.1
nbformat @ file:///home/conda/feedstock_root/build_artifacts/nbformat_1594060262917/work
notebook @ file:///D:/bld/notebook_1594079006001/work
numpy==1.16.6
oauthlib==3.1.0
opencv-python==4.3.0.36
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1589925210001/work
pandas==1.0.5
pandocfilters==1.4.2
parso==0.7.0
pickleshare==0.7.5
Pillow==7.2.0
portalocker==1.7.0
prometheus-client @ file:///home/conda/feedstock_root/build_artifacts/prometheus_client_1590412252446/work
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1592500439797/work
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.6.1
pylint==2.5.3
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2020.1
pywin32==228
pywinpty==0.5.7
pyzmq==19.0.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scipy==1.5.1
Send2Trash==1.5.0
six @ file:///home/conda/feedstock_root/build_artifacts/six_1590081179328/work
tensorboard==2.2.2
tensorboard-plugin-wit==1.7.0
terminado==0.8.3
testpath==0.4.4
toml==0.10.1
tornado==6.0.4
tqdm==4.47.0
traitlets==4.3.3
typed-ast==1.4.1
urllib3==1.22
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1592931742287/work
webencodings==0.5.1
Werkzeug==1.0.1
wincertstore==0.2
zipp==3.1.0
